bruh

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/utils/DcMotorUtils.java
@@ -7,7 +7,7 @@ public class DcMotorUtils {
     // Except it only sets the power and you have to run it in the loop() thing, because yes
     // NOTE: in the future add lambdas so that you can pass your preferred decel func to it
     static public void moveByTicks(DcMotor motor, double desiredTicks, double startDec, double tolerance) {
-        int diff = (int) getDifference(motor, desiredTicks);
+        double diff = getDifference(motor, desiredTicks);
         // Deceleration: min(x * 1/startDec, 1) where x = difference
         if (Math.abs(diff) > tolerance ) {
             int direction = (int) (diff/Math.abs(diff));
